dbinspector tweaks
